refactor: 全角・半角に関する処理を`katakana_english.py`に移動
