+++ voicevox_engine/tts_pipeline/katakana_english.py
@@ -1,10 +1,12 @@
 """英単語をカタカナ読みにする処理"""
 
 import re
+from typing import NewType, TypeGuard
 
 import e2k
 
-from ..utility.character_width_utility import HankakuAlphabet, is_hankaku_alphabet
+# 半角アルファベット文字列を示す型
+HankakuAlphabet = NewType("HankakuAlphabet", str)
 
 _global_c2k: e2k.C2K | None = None
 
@@ -47,6 +49,22 @@ ojt_alphabet_kana_mapping = {
     "Y": "ワイ",
     "Z": "ズィー",
 }
+
+
+def is_hankaku_alphabet(text: str) -> TypeGuard[HankakuAlphabet]:
+    """文字列が半角アルファベットのみで構成されているかを判定する"""
+    return bool(re.fullmatch("[a-zA-Z]+", text))
+
+
+def convert_zenkaku_alphabet_to_hankaku(text: str) -> str:
+    """全角アルファベットを半角に変換する"""
+    # TODO: ユーザー辞書にも似た関数があるため、共通化を検討する
+    return text.translate(
+        str.maketrans(
+            "".join(chr(0xFF01 + i) for i in range(94)),
+            "".join(chr(0x21 + i) for i in range(94)),
+        )
+    )
 
 
 def should_convert_english_to_katakana(string: HankakuAlphabet) -> bool:
